+++ src/game/terrain/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod flora;
 pub(crate) mod pond;
 pub(crate) mod tree;
 pub(crate) mod utils;
+pub(crate) mod wind_sway;